Update script.js
So I started writing more code on to this file today and started to realize a few things. 1) The code was starting to get very confusing and I felt like I was getting lost in my own code. 2) It was way too complex for it being a 'Simple Project' and so I decided to scrap it and start over with the same premise as before, make a simple To-Do application and give it some basic functionality. 

I went ahead and gave myself the opportunity to add in a filter with colors corresponding to the given filter values later on, but again I wonder how far I should take this before it gets beyond the scope of the goal of this folder which is practice.